+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Reports for timed-out tests and hooks.
+- Option `WithGroupHooks` to configure `BeforeAll` & `AfterAll` hooks groupping in reports.
+- Option `WithTimeouts` to disable timeout handling.
+
 ## [1.2.2] - 2026-06-18
 
 ### Fixed